# Api/doc/openapi.json
{"openapi": "3.1.0", "info": {"title": "AREA-API", "version": "1.0.0"}, "paths": {"/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/about.json": {"get": {"tags": ["Default"], "summary": "About", "description": "About", "operationId": "about_about_json_get", "responses": {"200": {"description": "About", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/About"}}}}}}}, "/client.apk": {"get": {"tags": ["Default"], "summary": "Client apk", "description": "Client apk", "operationId": "client_client_apk_get", "responses": {"200": {"description": "Client"}}}}, "/flux": {"post": {"tags": ["Flux"], "summary": "Create/Modify flux", "description": "Create/Modify flux", "operationId": "create_flux_flux_post", "parameters": [{"name": "verify", "in": "query", "required": false, "schema": {"type": "boolean", "default": false, "title": "Verify"}}, {"name": "access-token", "in": "header", "required": false, "schema": {"type": "string", "title": "Access-Token"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FluxCreateOrModify-Input"}}}}, "responses": {"201": {"description": "Flux", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FluxSend"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"example": {"detail": "string"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"example": {"detail": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/flux/fluxs": {"get": {"tags": ["Flux"], "summary": "Get all fluxs", "description": "Get all fluxs", "operationId": "get_fluxs_flux_fluxs_get", "parameters": [{"name": "access-token", "in": "header", "required": false, "schema": {"type": "string", "title": "Access-Token"}}], "responses": {"200": {"description": "List of fluxs", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/FluxBasicData"}, "title": "Response Get Fluxs Flux Fluxs Get"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"example": {"detail": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/flux/{fluxId}": {"get": {"tags": ["Flux"], "summary": "Get flux", "description": "Get flux", "operationId": "get_flux_flux__fluxId__get", "parameters": [{"name": "fluxId", "in": "path", "required": true, "schema": {"type": "integer", "title": "Fluxid"}}, {"name": "access-token", "in": "header", "required": false, "schema": {"type": "string", "title": "Access-Token"}}], "responses": {"200": {"description": "Flux", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FluxCreateOrModify-Output"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"example": {"detail": "string"}}}}, "404": {"description": "Flux not found", "content": {"application/json": {"example": {"detail": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/user/signup": {"post": {"tags": ["User"], "summary": "Create user", "description": "Create user", "operationId": "createUser_user_signup_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}, "required": true}, "responses": {"201": {"description": "User created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserToken"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"example": {"detail": "string"}}}}, "409": {"description": "Conflict", "content": {"application/json": {"example": {"detail": "string"}}}}, "401": {"description": "Issue when creating authorization token", "content": {"application/json": {"example": {"detail": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/user/login": {"post": {"tags": ["User"], "summary": "Login user", "description": "Login user", "operationId": "loginUser_user_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserLogin"}}}, "required": true}, "responses": {"200": {"description": "User logged", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserToken"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"example": {"detail": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/user/me": {"get": {"tags": ["User"], "summary": "Get user", "description": "Get user", "operationId": "getUser_user_me_get", "parameters": [{"name": "access-token", "in": "header", "required": false, "schema": {"type": "string", "title": "Access-Token"}}], "responses": {"200": {"description": "User", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserMe"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"example": {"detail": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/services/services": {"get": {"tags": ["Services"], "summary": "Services", "description": "Get services", "operationId": "get_services_services_services_get", "parameters": [{"name": "access-token", "in": "header", "required": false, "schema": {"type": "string", "title": "Access-Token"}}], "responses": {"200": {"description": "Services", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ServicesType"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"example": {"detail": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/services/metadata/{areaId}": {"get": {"tags": ["Services"], "summary": "Get action or reaction input schema", "description": "Get action or reaction input schema", "operationId": "get_action_or_reaction_services_metadata__areaId__get", "parameters": [{"name": "areaId", "in": "path", "required": true, "schema": {"type": "string", "title": "Areaid"}}, {"name": "access-token", "in": "header", "required": false, "schema": {"type": "string", "title": "Access-Token"}}], "responses": {"200": {"description": "Action or reaction schema", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ServiceMetadataSend"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"example": {"detail": "string"}}}}, "404": {"description": "Service not found", "content": {"application/json": {"example": {"detail": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/services/{service}/authorize_url": {"get": {"tags": ["Services"], "summary": "Authorization URL", "description": "Get authorization url", "operationId": "authorize_url_services__service__authorize_url_get", "parameters": [{"name": "service", "in": "path", "required": true, "schema": {"type": "string", "title": "Service"}}, {"name": "access-token", "in": "header", "required": false, "schema": {"type": "string", "title": "Access-Token"}}], "responses": {"200": {"description": "Authorization URL", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthorizationUrl"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"example": {"detail": "string"}}}}, "404": {"description": "Service not found", "content": {"application/json": {"example": {"detail": "string"}}}}, "409": {"description": "Service already exist", "content": {"application/json": {"example": {"detail": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/services/{service}/authorize": {"get": {"tags": ["Services"], "summary": "Authorize", "description": "Authorize", "operationId": "authorize_services__service__authorize_get", "parameters": [{"name": "service", "in": "path", "required": true, "schema": {"type": "string", "title": "Service"}}, {"name": "state", "in": "query", "required": true, "schema": {"type": "string", "title": "State"}}, {"name": "code", "in": "query", "required": true, "schema": {"type": "string", "title": "Code"}}, {"name": "scope", "in": "query", "required": true, "schema": {"type": "string", "title": "Scope"}}, {"name": "error", "in": "query", "required": false, "schema": {"type": "string", "title": "Error"}}], "responses": {"200": {"description": "Authorize", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/services/{service}/is_connected": {"get": {"tags": ["Services"], "summary": "Is connected", "description": "Is connected", "operationId": "is_connected_services__service__is_connected_get", "parameters": [{"name": "service", "in": "path", "required": true, "schema": {"type": "string", "title": "Service"}}, {"name": "access-token", "in": "header", "required": false, "schema": {"type": "string", "title": "Access-Token"}}], "responses": {"200": {"description": "Is connected", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/services/{service}/{serviceType}": {"get": {"tags": ["Services"], "summary": "Get all actions or reactions", "description": "Get all actions or reactions", "operationId": "get_actions_or_reactions_services__service___serviceType__get", "parameters": [{"name": "service", "in": "path", "required": true, "schema": {"type": "string", "title": "Service"}}, {"name": "serviceType", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/ServiceType"}}, {"name": "access-token", "in": "header", "required": false, "schema": {"type": "string", "title": "Access-Token"}}], "responses": {"200": {"description": "List of actions or reactions", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/AREADataSmall"}, "title": "Response Get Actions Or Reactions Services  Service   Servicetype  Get"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"example": {"detail": "string"}}}}, "404": {"description": "Service not found", "content": {"application/json": {"example": {"detail": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/services/test": {"post": {"tags": ["Services"], "summary": "Test action or reaction", "description": "Test action or reaction", "operationId": "test_action_or_reaction_services_test_post", "parameters": [{"name": "access-token", "in": "header", "required": false, "schema": {"type": "string", "title": "Access-Token"}}], "responses": {"200": {"description": "Test action or reaction", "content": {"application/json": {"schema": {}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"example": {"detail": "string"}}}}, "404": {"description": "Service not found", "content": {"application/json": {"example": {"detail": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"AREADataSmall": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}}, "type": "object", "required": ["id", "name", "description"], "title": "AREADataSmall", "example": {"description": "???", "id": "???", "name": "???"}}, "About": {"properties": {"client": {"$ref": "#/components/schemas/Client"}, "server": {"$ref": "#/components/schemas/Server"}}, "type": "object", "required": ["client", "server"], "title": "About"}, "Action": {"properties": {"name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}}, "type": "object", "required": ["name", "description"], "title": "Action"}, "AuthorizationUrl": {"properties": {"url": {"type": "string", "title": "Url"}}, "type": "object", "required": ["url"], "title": "AuthorizationUrl", "description": "Authorization URL Model", "example": {"url": "https://accounts.google.com/o/oauth2/auth?response_type=???&client_id=???&redirect_uri=???&scope=???&state=???&access_type=???&include_granted_scopes=???"}}, "Client": {"properties": {"host": {"type": "string", "title": "Host"}}, "type": "object", "required": ["host"], "title": "Client"}, "DataConfigurationType": {"type": "string", "enum": ["text", "number", "textMultiline", "select", "checkbox", "date"], "title": "DataConfigurationType"}, "FluxBasicData": {"properties": {"id": {"type": "integer", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "active": {"type": "boolean", "title": "Active"}}, "type": "object", "required": ["id", "name", "description", "active"], "title": "FluxBasicData", "description": "Flux Basic Data Model"}, "FluxCreateOrModify-Input": {"properties": {"id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}, "name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "nodes": {"items": {"$ref": "#/components/schemas/FluxNode-Input"}, "type": "array", "title": "Nodes"}, "edges": {"items": {"$ref": "#/components/schemas/FluxEdge"}, "type": "array", "title": "Edges"}}, "type": "object", "required": ["name", "description", "nodes", "edges"], "title": "FluxCreateOrModify", "description": "Flux Create Model"}, "FluxCreateOrModify-Output": {"properties": {"id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}, "name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "nodes": {"items": {"$ref": "#/components/schemas/FluxNode-Output"}, "type": "array", "title": "Nodes"}, "edges": {"items": {"$ref": "#/components/schemas/FluxEdge"}, "type": "array", "title": "Edges"}}, "type": "object", "required": ["id", "name", "description", "nodes", "edges"], "title": "FluxCreateOrModify", "description": "Flux Create Model"}, "FluxEdge": {"properties": {"id": {"type": "string", "title": "Id"}, "nodeStartId": {"type": "string", "title": "Nodestartid"}, "nodeEndId": {"type": "string", "title": "Nodeendid"}, "inputIndex": {"type": "integer", "title": "Inputindex"}, "outputIndex": {"type": "integer", "title": "Outputindex"}, "prevStartPos": {"$ref": "#/components/schemas/FluxPos"}, "currStartPos": {"$ref": "#/components/schemas/FluxPos"}, "prevEndPos": {"$ref": "#/components/schemas/FluxPos"}, "currEndPos": {"$ref": "#/components/schemas/FluxPos"}}, "type": "object", "required": ["id", "nodeStartId", "nodeEndId", "inputIndex", "outputIndex", "prevStartPos", "currStartPos", "prevEndPos", "currEndPos"], "title": "FluxEdge", "description": "Flux Edge Model"}, "FluxInputData-Input": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "data": {"items": {"type": "string"}, "type": "array", "title": "Data"}, "inputType": {"type": "string", "title": "Inputtype"}, "type": {"type": "string", "title": "Type"}, "required": {"type": "boolean", "title": "Required"}, "value": {"anyOf": [{}, {"type": "null"}], "title": "Value"}}, "type": "object", "required": ["id", "name", "data", "inputType", "type", "required"], "title": "FluxInputData", "description": "Flux Input Data Model"}, "FluxInputData-Output": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "data": {"items": {"type": "string"}, "type": "array", "title": "Data"}, "inputType": {"type": "string", "title": "Inputtype"}, "type": {"type": "string", "title": "Type"}, "required": {"type": "boolean", "title": "Required"}, "value": {"anyOf": [{}, {"type": "null"}], "title": "Value"}}, "type": "object", "required": ["id", "name", "data", "inputType", "type", "required", "value"], "title": "FluxInputData", "description": "Flux Input Data Model"}, "FluxNode-Input": {"properties": {"id": {"type": "string", "title": "Id"}, "numberInputs": {"anyOf": [{"type": "integer"}, {"type": "string"}], "title": "Numberinputs"}, "numberOutputs": {"anyOf": [{"type": "integer"}, {"type": "string"}], "title": "Numberoutputs"}, "service": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Service"}, "subService": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Subservice"}, "subServiceId": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Subserviceid"}, "inputDataFromSubServiceId": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Inputdatafromsubserviceid"}, "inputsData": {"anyOf": [{"items": {"$ref": "#/components/schemas/FluxInputData-Input"}, "type": "array"}, {"type": "null"}], "title": "Inputsdata", "default": []}, "prevPosition": {"$ref": "#/components/schemas/FluxPos"}, "currPosition": {"$ref": "#/components/schemas/FluxPos"}, "type": {"type": "string", "title": "Type"}, "inputEdgeIds": {"items": {"type": "string"}, "type": "array", "title": "Inputedgeids"}, "outputEdgeIds": {"items": {"type": "string"}, "type": "array", "title": "Outputedgeids"}, "title": {"type": "string", "title": "Title"}, "img": {"type": "string", "title": "Img"}}, "type": "object", "required": ["id", "numberInputs", "numberOutputs", "prevPosition", "currPosition", "type", "inputEdgeIds", "outputEdgeIds", "title", "img"], "title": "FluxNode", "description": "Flux Node Model"}, "FluxNode-Output": {"properties": {"id": {"type": "string", "title": "Id"}, "numberInputs": {"anyOf": [{"type": "integer"}, {"type": "string"}], "title": "Numberinputs"}, "numberOutputs": {"anyOf": [{"type": "integer"}, {"type": "string"}], "title": "Numberoutputs"}, "service": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Service"}, "subService": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Subservice"}, "subServiceId": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Subserviceid"}, "inputDataFromSubServiceId": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Inputdatafromsubserviceid"}, "inputsData": {"anyOf": [{"items": {"$ref": "#/components/schemas/FluxInputData-Output"}, "type": "array"}, {"type": "null"}], "title": "Inputsdata", "default": []}, "prevPosition": {"$ref": "#/components/schemas/FluxPos"}, "currPosition": {"$ref": "#/components/schemas/FluxPos"}, "type": {"type": "string", "title": "Type"}, "inputEdgeIds": {"items": {"type": "string"}, "type": "array", "title": "Inputedgeids"}, "outputEdgeIds": {"items": {"type": "string"}, "type": "array", "title": "Outputedgeids"}, "title": {"type": "string", "title": "Title"}, "img": {"type": "string", "title": "Img"}}, "type": "object", "required": ["id", "numberInputs", "numberOutputs", "service", "subService", "subServiceId", "inputDataFromSubServiceId", "inputsData", "prevPosition", "currPosition", "type", "inputEdgeIds", "outputEdgeIds", "title", "img"], "title": "FluxNode", "description": "Flux Node Model"}, "FluxPos": {"properties": {"x": {"type": "number", "title": "X"}, "y": {"type": "number", "title": "Y"}}, "type": "object", "required": ["x", "y"], "title": "FluxPos", "description": "Flux Position Model"}, "FluxSend": {"properties": {"id": {"type": "integer", "title": "Id"}}, "type": "object", "required": ["id"], "title": "FluxSend", "description": "Flux Send Model"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Reaction": {"properties": {"name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}}, "type": "object", "required": ["name", "description"], "title": "Reaction"}, "Server": {"properties": {"current_time": {"type": "integer", "title": "Current Time"}, "services": {"items": {"$ref": "#/components/schemas/Service"}, "type": "array", "title": "Services"}}, "type": "object", "required": ["current_time", "services"], "title": "Server"}, "Service": {"properties": {"name": {"type": "string", "title": "Name"}, "actions": {"items": {"$ref": "#/components/schemas/Action"}, "type": "array", "title": "Actions"}, "reactions": {"items": {"$ref": "#/components/schemas/Reaction"}, "type": "array", "title": "Reactions"}}, "type": "object", "required": ["name", "actions", "reactions"], "title": "Service"}, "ServiceMetadataSend": {"properties": {"id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Id"}, "name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "type": {"$ref": "#/components/schemas/ServiceType"}, "inputsData": {"items": {"$ref": "#/components/schemas/inputData"}, "type": "array", "title": "Inputsdata", "default": []}, "outputsData": {"items": {"$ref": "#/components/schemas/outputData"}, "type": "array", "title": "Outputsdata", "default": []}}, "type": "object", "required": ["id", "name", "description", "type", "inputsData", "outputsData"], "title": "ServiceMetadataSend"}, "ServiceType": {"type": "string", "enum": ["action", "reaction"], "title": "ServiceType"}, "ServicesType": {"properties": {"action": {"items": {"type": "string"}, "type": "array", "title": "Action"}, "reaction": {"items": {"type": "string"}, "type": "array", "title": "Reaction"}}, "type": "object", "required": ["action", "reaction"], "title": "ServicesType"}, "UserCreate": {"properties": {"email": {"type": "string", "title": "Email"}, "password": {"type": "string", "title": "Password"}, "surname": {"type": "string", "title": "Surname"}, "name": {"type": "string", "title": "Name"}}, "type": "object", "required": ["email", "password", "surname", "name"], "title": "UserCreate", "description": "User Create Model", "example": {"email": "test@test.com", "name": "Test", "password": "Test1234", "surname": "Test"}}, "UserLogin": {"properties": {"email": {"type": "string", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "UserLogin", "description": "User Login Model", "example": {"email": "test@test.com", "password": "Test1234"}}, "UserMe": {"properties": {"id": {"type": "integer", "title": "Id"}, "email": {"type": "string", "title": "Email"}, "surname": {"type": "string", "title": "Surname"}, "name": {"type": "string", "title": "Name"}}, "type": "object", "required": ["id", "email", "surname", "name"], "title": "UserMe", "example": {"email": "test@test.com", "id": 1, "name": "Test", "surname": "Test"}}, "UserToken": {"properties": {"access_token": {"type": "string", "title": "Access Token"}}, "type": "object", "required": ["access_token"], "title": "UserToken", "description": "User Token Model"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "inputData": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "inputType": {"$ref": "#/components/schemas/DataConfigurationType"}, "type": {"type": "string", "title": "Type"}, "required": {"type": "boolean", "title": "Required", "default": false}, "data": {"items": {"type": "string"}, "type": "array", "title": "Data", "default": []}}, "type": "object", "required": ["id", "name", "inputType", "type", "required", "data"], "title": "inputData"}, "outputData": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "type": {"type": "string", "title": "Type"}}, "type": "object", "required": ["id", "name", "type"], "title": "outputData"}}}}